Adjusted pawn and king movement and capture
I adpeted in a similar way the movement and capture of the pawns and kings, in order to make it consistent. I found some bug and adjusted some of them, but other are still to be fixed. Such as the fact that a piece can be moved wherever you click  in certain circumstances. Probably some problem with the selection of a piece. I started also the capture system.