Fixed Horse Bug and Adjusted Click
I finally re-opened the project that I haven't touched for a while. In order to fix the horse bug and to adjust the click at the same time I had to Split the Cell.cs and Piece.cs in two toher scripts and to use another collider for both cells and pieces. In this way there are colliders that handles only the click and other (really small) that handle the position and capture of pieces.